+++ src/main/java/racingcar/generator/NumberGenerator.java
@@ -1,6 +1,6 @@
-package racingcar.util;
+package racingcar.generator;
 
 public interface NumberGenerator {
     int generateNumber(int min, int max);
-    
+
 }
