+++ src/main/java/racingcar/generator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
