+++ src/test/java/racingcar/generator/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.generator;
 
 import static org.assertj.core.api.Assertions.assertThat;
